feat: update example with new expo version
